requirements.txt
dj-database-url==0.4.1
Django==1.10.3
django-filter==1.0.0
djangorestframework==3.5.3
google-api-python-client==1.5.5
httplib2==0.9.2
oauth2client==4.0.0
oauthlib==2.0.0
six==1.10.0
uritemplate==3.0.0
